fix mainactivity
