+++ app/src/main/java/com/example/c0115334/operator/MainActivity.java
@@ -4,26 +4,31 @@ import android.app.AlarmManager;
 import android.app.PendingIntent;
 import android.content.Context;
 import android.content.Intent;
+import android.database.Cursor;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
 import android.app.TimePickerDialog;
 
+import java.util.ArrayList;
 import java.util.Calendar;
+import java.util.List;
 
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-import android.app.AlarmManager;
-import android.app.PendingIntent;
-import android.content.Intent;
 import android.util.Log;
-import android.view.View;
+import android.widget.EditText;
+import android.widget.ListView;
 import android.widget.TextView;
 import android.widget.Toast;
 import android.widget.TimePicker;
 
-public class MainActivity extends AppCompatActivity implements View.OnClickListener{
+public class MainActivity extends AppCompatActivity implements View.OnClickListener {
+    private List<TodoItem> list = new ArrayList<TodoItem>();
+    private ListView listview;
+    private TodoAdapter todo;
+    private Button btn_save;
+    private EditText txt1;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -32,6 +37,42 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
 
         /*Button betTimeSet = (Button)findViewById(R.id.button1);
         betTimeSet.setOnClickListener(this);*/
+        todo = new TodoAdapter(this);
+        Cursor c = todo.getAllList();
+        if (c.moveToFirst()) {
+            do {
+                TodoItem item = new TodoItem();
+                item.set_id(c.getInt(c.getColumnIndex("_id")));
+                item.setMemo(c.getString(c.getColumnIndex("memo")));
+                list.add(item);
+            } while (c.moveToNext());
+        }
+
+        // 紐付け
+        listview = (ListView) findViewById(R.id.listView);
+
+        // ArrayAdapterへ設定
+        TodoListAdapter rowAdapater = new TodoListAdapter(this, 0, list);
+
+        // リストビューへ設定
+        listview.setAdapter(rowAdapater);
+
+        btn_save = (Button) findViewById(R.id.button3);
+        txt1 = (EditText) findViewById(R.id.editText);
+
+        btn_save.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                // 値を保存する
+                todo.insert(txt1.getText().toString());
+
+                // 画面を閉じる
+                //finish();
+            }
+        });
+
+        // todo
+        todo = new TodoAdapter(this);
     }
 
     @Override
@@ -64,13 +105,13 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                     @Override
                     public void onTimeSet(TimePicker view, int hourOfDay, int minute) {
                         //受け取った時間と今の時間を比較して明日なのか今日なのかを判定
-                        if (Integer.parseInt(String.valueOf(hourOfDay)+String.valueOf(minute)) >=
-                                Integer.parseInt(String.valueOf(calendar.get(Calendar.HOUR_OF_DAY)) + String.valueOf(calendar.get(Calendar.MINUTE)))){
-                            calendar.set(Calendar.DATE, calendar.get(Calendar.DATE)+1);
+                        if (Integer.parseInt(String.valueOf(hourOfDay) + String.valueOf(minute)) >=
+                                Integer.parseInt(String.valueOf(calendar.get(Calendar.HOUR_OF_DAY)) + String.valueOf(calendar.get(Calendar.MINUTE)))) {
+                            calendar.set(Calendar.DATE, calendar.get(Calendar.DATE) + 1);
                         }
                         //受け取った値をセット
                         calendar.set(Calendar.HOUR_OF_DAY, hourOfDay);
-                        calendar.set(Calendar.MINUTE, minute );
+                        calendar.set(Calendar.MINUTE, minute);
                         //0秒で起動
                         calendar.set(Calendar.SECOND, 0);
                         Intent intent = new Intent(getApplicationContext(), AlarmBroadcastReceiver.class);
@@ -105,7 +146,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                         startTime.set(Calendar.SECOND, 0);
                         Intent bootIntent = new Intent(MainActivity.this, BedTimeReceiver.class);
                         PendingIntent alarmIntent = PendingIntent.getBroadcast(MainActivity.this, 0, bootIntent, PendingIntent.FLAG_CANCEL_CURRENT);
-                        AlarmManager alarm = (AlarmManager)getSystemService(Context.ALARM_SERVICE);
+                        AlarmManager alarm = (AlarmManager) getSystemService(Context.ALARM_SERVICE);
                         long alarmStartTime = startTime.getTimeInMillis();
                         alarm.set(AlarmManager.RTC_WAKEUP, alarmStartTime, alarmIntent);
                         Toast.makeText(MainActivity.this, "通知をセットしました！", Toast.LENGTH_SHORT).show();
@@ -114,6 +155,34 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                 }, hour, minute, true);
         timePickerDialog.show();
 
+    }
+
+    @Override
+    protected void onRestart() {
+
+        Log.d("activity", "onRestart");
+
+        list.clear();
+        Cursor c = todo.getAllList();
+        if (c.moveToFirst()) {
+            do {
+                TodoItem item = new TodoItem();
+                item.set_id(c.getInt(c.getColumnIndex("_id")));
+                item.setMemo(c.getString(c.getColumnIndex("memo")));
+                list.add(item);
+            } while (c.moveToNext());
+        }
+
+        // 紐付け
+        listview = (ListView) findViewById(R.id.listView);
+
+        // ArrayAdapterへ設定
+        TodoListAdapter rowAdapater = new TodoListAdapter(this, 0, list);
+
+        // リストビューへ設定
+        listview.setAdapter(rowAdapater);
+
+        super.onRestart();
     }
 
 }
